Fix the out of sync replies!
The original flush routine would decide that it had finished swallowing input as soon as it
_ever_ got "no bytes available" even if it was just in between a byte somewhere.

That was far from reliable.  Instead, it should keep flushing and swallowing until the line
has been idle for the "inter byte timeout" period.  If bytes keep appearing before that
timeout, the timeout needs to be reset!

This became instantly more reliable.  Given the speeds and timeouts in other
implementaitons, this can mean that we (slave AA) will still be flushing the original
request to slave BB when slave BB's response comes back over the line.  That's fine. We
don't care about their message anyway.  We simply rely on the master's inter _message_ delay
being a bit longer than 10ms :)